Version 1.0 debugged
Fixed the Version 1.0 file. In the game, the window was not updating, which results in showing the trail of images.
This happened cause in the main menu function and the main loop function, there wasn't a background to load but in main loop function, there was a function "redraw function" , which draws a the current game events, which also includes the background of the game.